Rename `admin.api-resources.v1` to `admin.api-resources.v2`
