+++ features/admin.api-resources.v2/components/api-resource-panes/authorization-api-resource.tsx
