+++ features/admin.api-resources.v2/components/api-resource-panes/index.ts
