+++ features/admin.api-resources.v2/components/wizard/add-api-resource-steps/add-api-resource-basic.tsx
