+++ features/admin.api-resources.v2/components/wizard/add-api-resource-steps/permission-mapping-list-item.tsx
