+++ features/admin.api-resources.v2/configs/ui.ts
