+++ features/admin.applications.v1/components/api-authorization/api-authorization.tsx
@@ -40,7 +40,7 @@ import { Dispatch } from "redux";
 import { Divider, Grid, Icon } from "semantic-ui-react";
 import { SubscribedAPIResources } from "./subscribed-api-resources";
 import { AuthorizeAPIResource } from "./wizard/authorize-api-resource";
-import { useAPIResources } from "../../../admin.api-resources.v1/api";
+import { useAPIResources } from "../../../admin.api-resources.v2/api";
 import { AppState, FeatureConfigInterface, history } from "../../../admin.core.v1";
 import {
     authorizeAPI,
