+++ features/admin.applications.v1/components/api-authorization/wizard/authorize-api-resource.tsx
@@ -40,10 +40,10 @@ import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
 import { DropdownItemProps, DropdownProps, Grid, Header, Label, Modal } from "semantic-ui-react";
-import { useAPIResources } from "../../../../admin.api-resources.v1/api";
-import { APIResourceCategories, APIResourcesConstants } from "../../../../admin.api-resources.v1/constants";
-import { APIResourceInterface, APIResourcePermissionInterface } from "../../../../admin.api-resources.v1/models";
-import { APIResourceUtils } from "../../../../admin.api-resources.v1/utils/api-resource-utils";
+import { useAPIResources } from "../../../../admin.api-resources.v2/api";
+import { APIResourceCategories, APIResourcesConstants } from "../../../../admin.api-resources.v2/constants";
+import { APIResourceInterface, APIResourcePermissionInterface } from "../../../../admin.api-resources.v2/models";
+import { APIResourceUtils } from "../../../../admin.api-resources.v2/utils/api-resource-utils";
 import useScopesOfAPIResources from "../../../api/use-scopes-of-api-resources";
 import { Policy, PolicyInfo, policyDetails } from "../../../constants/api-authorization";
 import { ApplicationTemplateIdTypes } from "../../../models";
