+++ features/admin.core.v1/configs/app.ts
@@ -21,7 +21,7 @@ import { DocumentationProviders, DocumentationStructureFileTypes } from "@wso2is
 import { I18nModuleInitOptions, I18nModuleOptionsInterface, MetaI18N, generateBackendPaths } from "@wso2is/i18n";
 import { getFeatureGateResourceEndpoints } from "../../admin.extensions.v1/components/feature-gate/configs";
 import { getExtendedFeatureResourceEndpoints } from "../../admin.extensions.v1/configs/endpoints";
-import { getAPIResourceEndpoints } from "../../admin.api-resources.v1/configs/endpoint";
+import { getAPIResourceEndpoints } from "../../admin.api-resources.v2/configs/endpoint";
 import { getApplicationsResourceEndpoints } from "../../admin.applications.v1/configs/endpoints";
 import isLegacyAuthzRuntime from "../../admin.authorization.v1/utils/get-legacy-authz-runtime";
 import { getBrandingResourceEndpoints } from "../../admin.branding.v1/configs/endpoints";
