+++ features/admin.extensions.v1/configs/common.tsx
@@ -41,7 +41,7 @@ export const commonConfig: CommonConfig = {
                 children: [
                     {
                         component: lazy(() =>
-                            import("../../admin.api-resources.v1/pages/api-resource-edit")
+                            import("../../admin.api-resources.v2/pages/api-resource-edit")
                         ),
                         exact: true,
                         id: "apiResources-edit",
@@ -52,7 +52,7 @@ export const commonConfig: CommonConfig = {
                     },
                     {
                         component: lazy(() =>
-                            import("../../admin.api-resources.v1/pages/api-resources-internal-list")
+                            import("../../admin.api-resources.v2/pages/api-resources-internal-list")
                         ),
                         exact: true,
                         id: "apiResources-list",
@@ -63,7 +63,7 @@ export const commonConfig: CommonConfig = {
                     }
                 ],
                 component: lazy(() =>
-                    import("../../admin.api-resources.v1/pages/api-resources")
+                    import("../../admin.api-resources.v2/pages/api-resources")
                 ),
                 exact: true,
                 icon: {
