+++ features/admin.roles.v2/components/edit-role/edit-role-permission.tsx
@@ -47,10 +47,10 @@ import { Dispatch } from "redux";
 import { DropdownItemProps, DropdownProps } from "semantic-ui-react";
 import { RenderChip } from "./edit-role-common/render-chip";
 import { RoleAPIResourcesListItem } from "./edit-role-common/role-api-resources-list-item";
-import { useAPIResources } from "../../../admin.api-resources.v1/api";
-import { useGetAuthorizedAPIList } from "../../../admin.api-resources.v1/api/useGetAuthorizedAPIList";
-import { APIResourceCategories, APIResourcesConstants } from "../../../admin.api-resources.v1/constants";
-import { APIResourceUtils } from "../../../admin.api-resources.v1/utils/api-resource-utils";
+import { useAPIResources } from "../../../admin.api-resources.v2/api";
+import { useGetAuthorizedAPIList } from "../../../admin.api-resources.v2/api/useGetAuthorizedAPIList";
+import { APIResourceCategories, APIResourcesConstants } from "../../../admin.api-resources.v2/constants";
+import { APIResourceUtils } from "../../../admin.api-resources.v2/utils/api-resource-utils";
 import { useGetCurrentOrganizationType } from "../../../admin.organizations.v1/hooks/use-get-organization-type";
 import { getAPIResourceDetailsBulk, updateRoleDetails, useAPIResourceDetails } from "../../api";
 import { RoleAudienceTypes } from "../../constants/role-constants";
