+++ features/admin.roles.v2/components/wizard-updated/application-role-wizard.tsx
@@ -37,9 +37,9 @@ import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
 import { DropdownItemProps, DropdownProps, Grid, Header, Label, Modal } from "semantic-ui-react";
 import { Policy } from "../../../admin.extensions.v1/components/application/constants";
-import { APIResourceCategories, APIResourcesConstants } from "../../../admin.api-resources.v1/constants";
-import { APIResourceInterface } from "../../../admin.api-resources.v1/models";
-import { APIResourceUtils } from "../../../admin.api-resources.v1/utils/api-resource-utils";
+import { APIResourceCategories, APIResourcesConstants } from "../../../admin.api-resources.v2/constants";
+import { APIResourceInterface } from "../../../admin.api-resources.v2/models";
+import { APIResourceUtils } from "../../../admin.api-resources.v2/utils/api-resource-utils";
 import useSubscribedAPIResources from "../../../admin.applications.v1/api/use-subscribed-api-resources";
 import { AuthorizedAPIListItemInterface } from "../../../admin.applications.v1/models/api-authorization";
 import { ApplicationInterface } from "../../../admin.applications.v1/models/application";
