+++ features/admin.roles.v2/components/wizard-updated/role-permissions/role-permissions.tsx
@@ -35,9 +35,9 @@ import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
 import { DropdownItemProps, DropdownProps, Header, Label } from "semantic-ui-react";
 import { RoleAPIResourcesListItem } from "./components/role-api-resources-list-item";
-import { useAPIResources } from "../../../../admin.api-resources.v1/api";
-import { APIResourceCategories, APIResourcesConstants } from "../../../../admin.api-resources.v1/constants";
-import { APIResourceUtils } from "../../../../admin.api-resources.v1/utils/api-resource-utils";
+import { useAPIResources } from "../../../../admin.api-resources.v2/api";
+import { APIResourceCategories, APIResourcesConstants } from "../../../../admin.api-resources.v2/constants";
+import { APIResourceUtils } from "../../../../admin.api-resources.v2/utils/api-resource-utils";
 import { Policy } from "../../../../admin.applications.v1/constants/api-authorization";
 import { useAPIResourceDetails, useGetAuthorizedAPIList } from "../../../api";
 import { RoleAudienceTypes } from "../../../constants/role-constants";
